refac(halo2): move `blake2b_writer_impl.h` to `vendors/halo2/src`
This is because `blake2b_writer_impl.h` tries to include tachyon
internal headers.
