+++ vendors/halo2/src/blake2b_writer_impl.h
@@ -1,5 +1,5 @@
-#ifndef VENDORS_HALO2_INCLUDE_BLAKE2B_WRITER_IMPL_H_
-#define VENDORS_HALO2_INCLUDE_BLAKE2B_WRITER_IMPL_H_
+#ifndef VENDORS_HALO2_SRC_BLAKE2B_WRITER_IMPL_H_
+#define VENDORS_HALO2_SRC_BLAKE2B_WRITER_IMPL_H_
 
 #include "rust/cxx.h"
 
@@ -30,4 +30,4 @@ class Blake2bWriterImpl {
 
 }  // namespace tachyon::halo2_api
 
-#endif  // VENDORS_HALO2_INCLUDE_BLAKE2B_WRITER_IMPL_H_
+#endif  // VENDORS_HALO2_SRC_BLAKE2B_WRITER_IMPL_H_
